Rename achievement directory to app
