+++ .gitignore
@@ -1,5 +1,5 @@
 .idea/
-achievement/public/
+public/
 db/config.json
 *.sublime-project
 *.sublime-workspace
